# contracts/nexus-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nexus AI Agent Office API",
    "description": "Backend API for the AI Agent Office OS SaaS Platform",
    "version": "v1"
  },
  "paths": {
...
    "/api/Integrations": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationConnectionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationConnectionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationConnectionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Integrations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntegrationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntegrationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntegrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Integrations/{id}": {
      "put": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIntegrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationConnectionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Integrations/google/auth-url": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "scopes",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "ads,analytics,search_console"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAuthUrlResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAuthUrlResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAuthUrlResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Integrations/google/callback": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Integrations/ads/campaigns": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "dateRange",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "LAST_30_DAYS"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Integrations/analytics/summary": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "dateRange",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "LAST_30_DAYS"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Integrations/analytics/dashboard": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "parameters": [
          {
            "name": "dateRange",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "30daysAgo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Integrations/canva/token": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CanvaTokenStoreResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanvaTokenStoreResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanvaTokenStoreResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Integrations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanvaTokenStoreRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CanvaTokenStoreRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CanvaTokenStoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanvaTokenStoreRequest": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "scope": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CanvaTokenStoreResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "tokenType": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "scope": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateIntegrationRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/IntegrationProvider"
          },
          "accountId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "scopes": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "GoogleAuthUrlResponse": {
        "type": "object",
        "properties": {
          "authUrl": {
            "type": "string"
          },
          "requestedScopes": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "IntegrationConnectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "$ref": "#/components/schemas/IntegrationProvider"
          },
          "accountId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/IntegrationStatus"
          },
          "scopes": {
            "type": "string"
          },
          "tokenExpiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastHealthCheck": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "IntegrationProvider": {
        "enum": [
          "GoogleBusiness",
          "Instagram",
          "GoogleAds",
          "Facebook",
          "SearchConsole",
          "GoogleAnalytics",
          "WhatsAppBusiness",
          "Canva"
        ],
        "type": "string"
      },
      "IntegrationStatus": {
        "enum": [
          "Connected",
          "Expired",
          "Error",
          "Disconnected"
        ],
        "type": "string"
      },
...
      "UpdateIntegrationRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}